pyproject.toml
[project]
name = "repo-lens"
version = "0.1.0"
description = "A powerful Git-aware repository assistant powered by LiteLLM and smolagents."
authors = [{ name = "Satish Jasthi", email = "satish.jasthi@example.com" }]
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.12",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "litellm>=1.49.0",
  "smolagents",
]

[project.scripts]
repo-lens = "repo_lens.cli:app"

[tool.uv]
dev-dependencies = [
  "pytest>=8.3",
  "pytest-mock>=3.12",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
